+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.6 (pending)
 - Support specification of location to post_photo, post_video
+- Proxy support (alpha)
 
 ## 1.0.5
 - New disable/enable media comments endpoints